Reorganize CLI commands into subpackages
