+++ src/valkyrie/cli/agent/__init__.py
@@ -1,0 +1,4 @@
+from valkyrie.cli.agent.group import agent
+from valkyrie.cli.agent.commands import agent_remove, download, install, list_installed_agents, push
+
+__all__ = ["agent", "agent_remove", "download", "install", "list_installed_agents", "push"]
