+++ src/valkyrie/cli/agent/group.py
@@ -1,0 +1,7 @@
+import click
+
+
+@click.group()
+def agent():
+    """Agent command group"""
+    pass
