+++ src/valkyrie/cli/benchmark/__init__.py
@@ -1,0 +1,4 @@
+from valkyrie.cli.benchmark.group import benchmark
+from valkyrie.cli.benchmark.tasks import tasks
+
+__all__ = ["benchmark", "tasks"]
