+++ src/valkyrie/cli/benchmark/group.py
@@ -1,0 +1,7 @@
+import click
+
+
+@click.group()
+def benchmark():
+    """Benchmark command group"""
+    pass
