+++ src/valkyrie/cli/config/group.py
@@ -1,0 +1,7 @@
+import click
+
+
+@click.group()
+def config():
+    """Config command group"""
+    pass
