+++ src/valkyrie/cli/run/__init__.py
@@ -1,0 +1,23 @@
+from valkyrie.cli.run.group import run
+from valkyrie.cli.run.start import start
+from valkyrie.cli.run.fetch import fetch
+from valkyrie.cli.run.results import results
+from valkyrie.cli.run.stop import stop
+from valkyrie.cli.run.analyze import analyze
+from valkyrie.cli.run.resume import resume, retry_command
+from valkyrie.cli.run.list import list_benchmarks
+from valkyrie.cli.run.outputs import output_path, outputs
+
+__all__ = [
+    "analyze",
+    "fetch",
+    "list_benchmarks",
+    "output_path",
+    "outputs",
+    "results",
+    "resume",
+    "retry_command",
+    "run",
+    "start",
+    "stop",
+]
