+++ src/valkyrie/cli/run/group.py
@@ -1,0 +1,7 @@
+import click
+
+
+@click.group()
+def run():
+    """Run command group"""
+    pass
